+++ src/dotnet/HQ.Abstractions/Emails/BaseEmail.cs
@@ -2,4 +2,5 @@ namespace HQ.Abstractions.Emails;
 
 public abstract class BaseEmail
 {
+    public Uri WebUrl { get; set; } = null!;
 }