+++ src/dotnet/HQ.Server/Services/EmailMessageService.cs
@@ -35,6 +35,8 @@ namespace HQ.Server.Services
 
         public async Task<Result> SendEmail<T>(EmailMessage emailMessage, T model, string to, string subject, MailPriority priority = MailPriority.Normal, IEnumerable<Attachment>? attachments = null, CancellationToken ct = default) where T : BaseEmail
         {
+            model.WebUrl = _options.CurrentValue.WebUrl;
+
             var request = new GetEmailTemplateV1.Request<T>()
             {
                 EmailMessage = emailMessage,
@@ -180,7 +182,27 @@ namespace HQ.Server.Services
 
             await SendEmail(EmailMessage.Notification, model, staff.Email, "[HQ] Time Submission Reminder", MailPriority.High, null, ct);
         }
+        public async Task SendPlanSubmissionReminderEmail(Guid staffId, DateOnly date, CancellationToken ct)
+        {
+            var staff = await _context.Staff
+                .AsNoTracking()
+                .SingleOrDefaultAsync(t => t.Id == staffId, ct);
 
+            if (staff == null || String.IsNullOrEmpty(staff.Email))
+            {
+                return;
+            }
+
+            var model = new NotificationEmail()
+            {
+                Heading = "Plan Reminder",
+                Message = $"Please remember to fill out your plan and update your status in HQ.",
+                ButtonLabel = "Open HQ",
+                ButtonUrl = _options.CurrentValue.WebUrl
+            };
+
+            await SendEmail(EmailMessage.Notification, model, staff.Email, "[HQ] Plan Submission Reminder", MailPriority.High, null, ct);
+        }
         public async Task SendRejectedTimeSubmissionReminderEmail(Guid staffId, DateOnly from, DateOnly to, CancellationToken ct)
         {
             var staff = await _context.Staff
@@ -201,6 +223,27 @@ namespace HQ.Server.Services
             };
 
             await SendEmail(EmailMessage.Notification, model, staff.Email, "[HQ] Rejected Time Reminder", MailPriority.High, null, ct);
+        }
+        public async Task SendPointSubmissionReminderEmail(Guid staffId, DateOnly from, DateOnly to, CancellationToken ct)
+        {
+            var staff = await _context.Staff
+                .AsNoTracking()
+                .SingleOrDefaultAsync(t => t.Id == staffId, ct);
+
+            if (staff == null || String.IsNullOrEmpty(staff.Email))
+            {
+                return;
+            }
+
+            var model = new NotificationEmail()
+            {
+                Heading = "Points Reminder",
+                Message = $"Please remember to fill out your planning points in HQ.",
+                ButtonLabel = "Open HQ",
+                ButtonUrl = _options.CurrentValue.WebUrl
+            };
+
+            await SendEmail(EmailMessage.Notification, model, staff.Email, "[HQ] Planning Point Submission Reminder", MailPriority.High, null, ct);
         }
     }
 }